Don't lose first messages when setting up network

+++ common/src/node/modules/messages.rs
@@ -49,7 +49,7 @@ impl fmt::Debug for NodeMessage {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.write_str(&format!(
             "NodeMessage( id: {}, msg: {:?}",
-            self.id.short(),
+            self.id,
             self.msg
         ))
     }
